requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.29.1
botocore==1.32.1
cachetools==5.3.0
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==39.0.2
defusedxml==0.7.1
Django==4.0.5
django-allauth==0.51.0
django-cors-headers==3.14.0
django-filter==22.1
django-rest-knox==4.2.0
django-use-email-as-username==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
google-api-core==2.11.0
google-auth==2.17.2
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
jmespath==1.0.1
Markdown==3.4.1
mysqlclient==2.1.1
numpy==1.24.4
oauthlib==3.2.2
Pillow==10.1.0
protobuf==4.22.1
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.7.0
shapely==2.0.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.15
whitenoise==6.4.0
wincertstore==0.2
zipp==3.15.0
